refactor: consolidate validation logic into answers module
- Move validation functions from separate validation.rs to cli/answers.rs
- Inline ValidationError enum in answers module for better encapsulation
- Remove validation module from lib.rs exports
- Keep all validation-related functionality colocated with answer collection
- Maintain existing test coverage and API compatibility

This change improves code organization by consolidating related functionality
and reducing the number of top-level modules.
