+++ src/lib.rs
@@ -28,8 +28,5 @@ pub mod template;
 /// Configuration handling for Baker templates.
 pub mod config;
 
-/// Answer validators
-pub mod validation;
-
 /// Importing templates from a directory into renderer.
 pub mod import;
